refactor: split gig.service.ts and job.service.ts
