+++ packages/nextjs/services/graphql/fetchers/gig/index.ts
@@ -1,0 +1,1 @@
+export * from "./gig.service";
