+++ packages/nextjs/services/graphql/fetchers/job/index.ts
@@ -1,0 +1,1 @@
+export * from "./job.service";
